backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
psycopg2-binary>=2.9.5
aiosqlite>=0.18.0
websockets>=11.0.0
redis>=4.5.1
python-dotenv>=1.0.0
pytest>=7.3.1
httpx>=0.24.0
pytest-asyncio>=0.21.0
